FIX: store VisiCut path in inkscape extension, Closes #258

+++ distribute/distribute.sh
@@ -101,7 +101,8 @@ then
   cd ..
   # hide doc directory from checkinstall
   # mv doc doctmp
-  fakeroot checkinstall --fstrans --reset-uid --type debian --install=no -y --pkgname visicut --pkgversion $VERSION --arch all --pkglicense LGPL --pkggroup other --pkgsource "http://visicut.org" --pkgaltsource "https://github.com/t-oster/VisiCut" --pakdir distribute/ --maintainer "'Thomas Oster <thomas.oster@rwth-aachen.de>'" --requires "java-runtime,potrace" make install -e PREFIX=/usr > /dev/null || { echo "error"; exit 1; }
+  test -f /usr/bin/visicut && { echo "error: please first uninstall visicut"; exit 1; }
+  fakeroot checkinstall --fstrans --reset-uid --type debian --install=no -y --pkgname visicut --pkgversion $VERSION --arch all --pkglicense LGPL --pkggroup other --pkgsource "http://visicut.org" --pkgaltsource "https://github.com/t-oster/VisiCut" --pakdir distribute/ --maintainer "'Thomas Oster <thomas.oster@rwth-aachen.de>'" --requires "bash,java-runtime,potrace" make install -e PREFIX=/usr > /dev/null || { echo "error"; exit 1; }
   rm description-pak
   rm -rf doc-pak
   # mv doctmp doc
